+++ js/main.js
@@ -2,10 +2,7 @@ import './form.js';
 import {renderPicturesMiniature} from './photo.js';
 import {fetchData} from './api.js';
 import {showAlert} from './util.js';
-import {onImagesFiltersButtonClick} from './filter.js';
-
-const ImagesFilters = document.querySelector('.img-filters');
-const form = ImagesFilters.querySelector('.img-filters__form');
+import {addImageFilters} from './filter.js';
 
 // Получение данных с сервера;
 fetchData(
@@ -13,8 +10,7 @@ fetchData(
   'GET',
   (picturesData) => {
     renderPicturesMiniature(picturesData);
-    ImagesFilters.classList.remove('img-filters--inactive');
-    form.addEventListener('click', (evt) => onImagesFiltersButtonClick(evt, picturesData));
+    addImageFilters(picturesData);
   },
   (onFail) => showAlert(onFail),
 );
